Fix TiltedPhoques::GetPath() to persist the correct path
There were a couple of issues with the way it was doing it. The attempt to reference __ImageBase wasn't working, so it was effectively NULL. And NULL could stop working (and __ImageBase can change) once SkyrimSE.exe was loaded because the GetModuleFileName() spoofing in SkyrimTogether was not robust; it isn't possible to reliably determine if base STR is asking in the face of multiple different hook systems.